+++ inc/ajax-login.php
@@ -94,3 +94,4 @@ function checkout_tabs_wp_ml_ajax_login(): void {
 }
 
 
+