+++ inc/cpf-lock.php
@@ -125,3 +125,4 @@ add_filter('update_user_metadata', function ($check, $object_id, $meta_key, $met
 }, 10, 4);
 
 
+