+++ inc/geolocation/enqueue.php
@@ -48,3 +48,4 @@ add_action('wp_enqueue_scripts', function () {
 });
 
 
+